+++ TestSupport/EfHelpers/SqlServerHelpers.cs
@@ -38,11 +38,12 @@ namespace TestSupport.EfHelpers
         /// </summary>
         /// <typeparam name="T"></typeparam>
         /// <param name="callingClass">this should be this, i.e. the class you are in</param>
+        /// <param name="applyExtraOption">Optional: action that allows you to add extra options to the builder</param>
         /// <returns></returns>
-        public static DbContextOptions<T> CreateUniqueClassOptions<T>(this object callingClass)
+        public static DbContextOptions<T> CreateUniqueClassOptions<T>(this object callingClass, Action<DbContextOptionsBuilder<T>> applyExtraOption = null)
             where T : DbContext
         {
-            return CreateOptionWithDatabaseName<T>(callingClass).Options;
+            return CreateOptionWithDatabaseName<T>(callingClass, null, applyExtraOption).Options;
         }
 #endif
 
@@ -75,12 +76,14 @@ namespace TestSupport.EfHelpers
         /// <param name="callingClass">this should be this, i.e. the class you are in</param>
         /// <param name="efLog">This is a method that receives a LogOutput whenever EF Core logs something</param>
         /// <param name="logLevel">Optional: Sets the logLevel you want to capture. Defaults to Information</param>
+        /// <param name="applyExtraOption">Optional: action that allows you to add extra options to the builder</param>
         /// <returns></returns>
         public static DbContextOptions<T> CreateUniqueClassOptionsWithLogging<T>(this object callingClass,
-            Action<LogOutput> efLog, LogLevel logLevel = LogLevel.Information)
+            Action<LogOutput> efLog, LogLevel logLevel = LogLevel.Information, 
+            Action<DbContextOptionsBuilder<T>> applyExtraOption = null)
             where T : DbContext
         {
-            return CreateOptionWithDatabaseName<T>(callingClass)
+            return CreateOptionWithDatabaseName<T>(callingClass, null, applyExtraOption)
                 .UseLoggerFactory(new LoggerFactory(new[] { new MyLoggerProviderActionOut(efLog, logLevel) }))
                 .Options;
         }
@@ -110,12 +113,14 @@ namespace TestSupport.EfHelpers
         /// </summary>
         /// <typeparam name="T"></typeparam>
         /// <param name="callingClass">this should be this, i.e. the class you are in</param>
+        /// <param name="applyExtraOption">Optional: action that allows you to add extra options to the builder</param>
         /// <param name="callingMember">Do not use: this is filled in by compiler</param>
         /// <returns></returns>
         public static DbContextOptions<T> CreateUniqueMethodOptions<T>(this object callingClass,
+            Action<DbContextOptionsBuilder<T>> applyExtraOption = null,
             [CallerMemberName] string callingMember = "") where T : DbContext
         {
-            return CreateOptionWithDatabaseName<T>(callingClass, callingMember).Options;
+            return CreateOptionWithDatabaseName<T>(callingClass, callingMember, applyExtraOption).Options;
         }
 #endif
 
@@ -150,14 +155,16 @@ namespace TestSupport.EfHelpers
         /// <param name="callingClass">this should be this, i.e. the class you are in</param>
         /// <param name="efLog">This is a method that receives a LogOutput whenever EF Core logs something</param>
         /// <param name="logLevel">Optional: Sets the logLevel you want to capture. Defaults to Information</param>
+        /// <param name="applyExtraOption">Optional: action that allows you to add extra options to the builder</param>
         /// <param name="callingMember">Do not use: this is filled in by compiler</param>
         /// <returns></returns>
         public static DbContextOptions<T> CreateUniqueMethodOptionsWithLogging<T>(this object callingClass,
             Action<LogOutput> efLog,
             LogLevel logLevel = LogLevel.Information,
+            Action<DbContextOptionsBuilder<T>> applyExtraOption = null,
             [CallerMemberName] string callingMember = "") where T : DbContext
         {
-            return CreateOptionWithDatabaseName<T>(callingClass, callingMember)
+            return CreateOptionWithDatabaseName<T>(callingClass, callingMember,applyExtraOption)
                 .UseLoggerFactory(new LoggerFactory(new[] { new MyLoggerProviderActionOut(efLog, logLevel) }))
                 .Options;
         }
@@ -210,13 +217,14 @@ namespace TestSupport.EfHelpers
         }
 #elif NETSTANDARD2_1
         private static DbContextOptionsBuilder<T> CreateOptionWithDatabaseName<T>(object callingClass,
-            string callingMember = null)
+            string callingMember, Action<DbContextOptionsBuilder<T>> applyExtraOption)
             where T : DbContext
         {
             var connectionString = callingClass.GetUniqueDatabaseConnectionString(callingMember);
             var builder = new DbContextOptionsBuilder<T>();
             builder.UseSqlServer(connectionString);
             builder.ApplyOtherOptionSettings();
+            applyExtraOption?.Invoke(builder);
 
             return builder;
         }