+++ TestSupport/EfHelpers/TimeThings.cs
@@ -13,6 +13,7 @@ namespace TestSupport.EfHelpers
     public class TimeThings : IDisposable
     {
         private readonly Stopwatch _stopwatch ;
+        private Action<TimeThingResult> _funcToCall;
         private readonly ITestOutputHelper _output;
         private readonly string _message;
         private readonly int _numRuns;
@@ -20,12 +21,28 @@ namespace TestSupport.EfHelpers
         /// <summary>
         /// This will measure the time it took from this class being created to it being disposed
         /// </summary>
+        /// <param name="message"></param>
+        /// <param name="numRuns">Optional: if the timing covers multiple runs of something, then set numRuns to the number of runs and it will give you the average per run</param>
+        public TimeThings(Action<TimeThingResult> result, string message = "", int numRuns = 0)
+            : this(message, numRuns)
+        {
+            _funcToCall = result;
+        }
+
+        /// <summary>
+        /// This will measure the time it took from this class being created to it being disposed and writes out to xUnit ITestOutputHelper
+        /// </summary>
         /// <param name="output"></param>
         /// <param name="message"></param>
-        /// <param name="numRuns">Optional: if the timing covers multiple runs of soemthing, then set numRuns to the number of runs and it will give you the average per run</param>
+        /// <param name="numRuns">Optional: if the timing covers multiple runs of something, then set numRuns to the number of runs and it will give you the average per run</param>
         public TimeThings(ITestOutputHelper output, string message = "", int numRuns = 0)
+            : this(message, numRuns)
         {
             _output = output;
+        }
+
+        private TimeThings(string message = "", int numRuns = 0)
+        {
             _message = message;
             _numRuns = numRuns;
             _stopwatch = new Stopwatch();
@@ -33,26 +50,17 @@ namespace TestSupport.EfHelpers
         }
 
         /// <summary>
-        /// When disposed it will write to the test output the time it took from creation
+        /// When disposed it will return the result, either via a action or by an output
         /// </summary>
         public void Dispose()
         {
             _stopwatch.Stop();
             var timeMilliseconds = _stopwatch.ElapsedTicks * 1000.0 / Stopwatch.Frequency;
-            var prefix = _numRuns > 1 ? $"{_numRuns:#,###} x " : "";
-            var suffix = _numRuns > 1 ? $", ave. per run = {TimeScaled(timeMilliseconds/_numRuns)}" : "";
-            _output?.WriteLine($"{prefix}{_message} took {timeMilliseconds:#,###.00} ms.{suffix}");
+            var result = new TimeThingResult(timeMilliseconds, _numRuns, _message);
+            _funcToCall?.Invoke(result);
+            _output?.WriteLine(result.ToString());
         }
 
-        private string TimeScaled(double timeMilliseconds)
-        {
-            if (timeMilliseconds > 5 * 1000)
-                return $"{timeMilliseconds / 1000:F3} sec.";
-            if (timeMilliseconds > 5)
-                return $"{timeMilliseconds:#,###.00} ms.";
-            if (timeMilliseconds > 5 / 1000.0)
-                return $"{timeMilliseconds * 1000:#,###.00} us.";
-            return $"{timeMilliseconds * 1000_000:#,###.0} ns.";
-        }
+
     }
 }