+++ ReleaseNotes.md
@@ -9,11 +9,11 @@
 
 - Improve EfSchemaCompare by detecting table sharing and account for nullable properties - see [breaking change](https://docs.microsoft.com/en-us/ef/core/what-is-new/ef-core-3.0/breaking-changes#dependent-entities-sharing-the-table-with-the-principal-are-now-optional).
 - EF Core 3 - Owned Type class properties are always nullable.
-- Add version of SqliteInMemory and SqlServerHelpers to return DbContextOptionsBuilder<T>
-- Add version of TimeThings that calls an action to obtain the time etc. 
 
-## 3.1.2
+## 3.2.0
 
+- New Feature: (only EF Core 3): The extension methods fro creating options for In-memory sqlite options and SQL Server options now have an optional parameter that allows you to add extra options to the DbContextOptionsBuilder<T>.
+- New Feature: TimeThings has version that returns the result of the timing.
 - Issue 32: Missing primary key in database causes a exception.
 
 ## 3.1.1